+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Sparse tensor quantization mode
     - Features at duplicated coordinates will be averaged automatically with `quantization_mode=ME.SparseTensorQuantizationMode.UNWEIGHTED_AVERAGE`
 - `SparseTensor.slice()` slicing features on discrete coordinates to continuous coordinates
+- CoordsManager.getKernelMapGPU returns long type tensors (Issue #125)
 
 
 ## [0.4.2] - 2020-03-13